If -u and a single non-upgradable package are specified, exit with 2
If a single package is specified on the command-line, and the -u
option is specified, we assume that the user wants to know whether
that package is upgradable. If it is not upgradable, we exit with
code 2.

Note that even if our constructor was not a regex or a task, but
a group, the pattern might still match multiple packages if the
architecture contains * (such as apt:linux-*). As such, we need
to specifically check that the pattern does not contain * in order
to make sure that only a single package was specified.